Fix `prep` component!
