+++ azure/pipeline/lib/constants.py
@@ -36,3 +36,10 @@ CAT_NOM_COLS = [
 
 CAT_ORD_COLS = [
 ]
+
+ENDPOINT_NAME='wisconsin-bca-endpoint'
+MODEL_NAME='wisconsin-BCa-model'
+ENV_NAME='wisconsin-bca-env'
+
+CPU_CLUSTER_TARGET='cpu-cluster'
+EXPERIMENT_NAME='wisconsin-bca'
